server/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aioice==0.9.0
aiortc==1.10.1
aiosignal==1.3.2
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
attrs==25.1.0
av==13.1.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cryptography==44.0.1
ctranslate2==4.5.0
cycler==0.12.1
dnspython==2.7.0
enum34==1.1.10
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
gast==0.6.0
google-crc32c==1.6.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.13.0
halo==0.0.31
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
imageio==2.37.0
jax==0.5.1
jaxlib==0.5.1
Jinja2==3.1.6
joblib==1.4.2
keras==3.8.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
lightning-utilities==0.14.0
llvmlite==0.44.0
log-symbols==0.0.14
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mediapipe==0.10.21
ml-dtypes==0.4.1
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.1.0
namex==0.0.8
networkx==3.4.2
numba==0.61.0
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.21.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
openwakeword==0.4.0
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pillow==11.1.0
propcache==0.3.0
protobuf==4.25.6
pvporcupine==1.9.5
PyAudio==0.2.14
pycparser==2.22
pyee==12.1.1
Pygments==2.19.1
pylibsrtp==0.11.0
pymongo==4.11.2
pyOpenSSL==25.0.0
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.5.0.post0
pytorch-triton-rocm==3.2.0+git4b3bb1f8
PyYAML==6.0.2
qasync==0.27.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
sentencepiece==0.2.0
setuptools==75.8.2
six==1.17.0
sounddevice==0.5.1
soundfile==0.13.1
spinners==0.0.24
sympy==1.13.3
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow-rocm==2.17.0
termcolor==2.5.0
threadpoolctl==3.6.0
tifffile==2025.2.18
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
urllib3==2.3.0
webrtcvad-wheels==2.0.14
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yarl==1.18.3
